requirements.txt
pytorch-lightning~=1.3.0
hydra-core~=1.0.0
wandb~=0.10.0
torch-scatter~=2.0.6
torch-sparse~=0.6.9
torch-cluster~=1.5.9
torch-spline-conv~=1.2.1
torch-geometric~=1.7.0
